Parallelized emcee!
In order to run multiple threads, lnProb and all its arguments need to be pickleable. priorFuncs is not pickleable since those are lambda functions not defined at the top level of the module, so pickle just saves their function name and when it unpickles it wouldn't be able to find the function definitions.

So as a workaround, I pass the priors array (which is pickleable) to lnProb which then calls interpretPriors to get priorFuncs. This step is slower than the profile evaluation (~0.5 ms compared to >~100 ms) so it's not much of a loss. Now runMCMC takes an nThreads argument and passes it to EnsembleSampler, so 8 threads can run in parallel on macbeth. Woohoo!